LayoutTests/platform/ipad/editing/selection/ios/show-selection-in-transformed-container-expected.txt



Apply transform
This test verifies that after focusing a visible input field in a body element that has been translated horizontally out of view, the caret is still visible. To run the test manually, first tap the input field and check that the caret shows up. Then, tap the button to apply the CSS transform, tap the input field again, and check that the caret is still visible.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS caretBefore.left is 18
PASS caretBefore.top is 12
PASS caretBefore.width is 2
PASS caretBefore.height is 24
FAIL caretAfter.left should be 18. Was 17.
PASS caretAfter.top is 12
FAIL caretAfter.width should be 2. Was 3.
PASS caretAfter.height is 24
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
